+++ src/Provider/ConfigServiceProvider.php
@@ -35,7 +35,7 @@ class ConfigServiceProvider implements ServiceProviderInterface
      */
     public function register(Application $app)
     {
-        $app['config'] = $app->share(function () use ($app) {
+        $app['config'] = $app->share(function () {
             switch (strtolower(pathinfo($this->filename, PATHINFO_EXTENSION))) {
                 case 'yml':
                     $config = (new YamlConfig(new YamlFileLoader(new FileLocator())))